requirements.txt
adbutils==0.3.4
allure-pytest==2.8.6
allure-python-commons==2.8.6
amqp==1.4.9
ansi2html==1.4.2
anyjson==0.3.3
apipkg==1.5
appdirs==1.4.3
Appium-Python-Client==0.26
asn1crypto==0.24.0
atomicwrites==1.1.5
attrs==18.1.0
Babel==2.7.0
backports.csv==1.0.7
bcrypt==3.1.7
beautifulsoup4==4.6.1
billiard==3.3.0.23
bs4==0.0.1
celery==3.1.26.post2
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
click==6.7
colorama==0.3.9
colorlog==4.0.2
concurrent-log-handler==0.9.9
configparser==5.0.0
cryptography==2.6.1
cryptokit==0.0.7
cssselect==1.0.3
ddt==1.1.2
decorator==4.4.0
defusedxml==0.5.0
demjson==2.2.4
deprecation==2.0.6
diff-match-patch==20181111
Django==2.0.3
django-celery==3.2.2
django-cors-headers==3.1.1
django-crispy-forms==1.8.0
django-filter==2.2.0
django-formtools==2.1
django-import-export==1.2.0
django-mdeditor==0.1.15
django-ranged-response==0.2.0
django-reversion==2.0.0
django-simple-captcha==0.5.10
django-stdimage==4.0.1
djangorestframework==3.10.3
docopt==0.6.2
dwebsocket==0.5.11
elasticsearch==7.7.0
enum34==1.1.6
et-xmlfile==1.0.1
eventlet==0.22.1
execnet==1.5.0
facebook-wda==0.3.4
fake-useragent==0.1.11
filetype==1.0.7
fire==0.1.3
Flask==1.0.2
flower==0.9.2
future==0.15.2
futures==3.1.1
gevent==1.3.6
greenlet==0.4.15
har2case==0.3.1
httpie==1.0.3
httplib2==0.9.2
HttpRunner==1.5.8
humanize==0.5.1
idna==2.6
importlib-metadata==1.2.0
itsdangerous==0.24
jdcal==1.4
Jinja2==2.10
jsonpath==0.82
jsonschema==3.2.0
kafka-python==1.4.6
kombu==3.0.37
locustio==0.11.0
logzero==1.5.0
lxml==4.2.5
Markdown==3.1.1
MarkupSafe==1.0
more-itertools==4.2.0
msgpack==0.5.6
mysqlclient==1.3.12
namedlist==1.7
nb-log==2.6
odfpy==1.4.0
openpyxl==2.5.9
packaging==19.0
paramiko==2.4.1
ParamUnittest==0.2
parse==1.11.1
pika==1.1.0
Pillow==5.2.0
pluggy==0.13.1
progress==1.5
progressbar2==3.39.3
py==1.5.4
pyasn1==0.4.7
pycparser==2.18
pycryptodome==3.8.2
pyee==5.0.0
Pygments==2.4.2
pyHook==1.5.1
pymongo==3.5.1
PyMySQL==0.9.3
PyNaCl==1.3.0
pyOpenSSL==17.5.0
pyparsing==2.4.0
pyppeteer==0.0.25
pyquery==1.4.0
pyrsistent==0.16.0
pytest==4.5.0
pytest-assume==2.2.1
pytest-forked==0.2
pytest-html==1.19.0
pytest-metadata==1.6.0
pytest-ordering==0.6
pytest-repeat==0.7.0
pytest-rerunfailures==8.0
pytest-xdist==1.23.2
PyTestReport==0.1.9.3
python-utils==2.3.0
pytz==2018.5
PyUserInput==0.1.11
pywin32==223
PyYAML==3.12
pyzmq==17.1.2
requests==2.22.0
requests-html==0.10.0
requests-toolbelt==0.8.0
retry==0.9.2
selenium==2.53.6
six==1.15.0
tablib==0.13.0
tomorrow==0.2.4
tomorrow3==1.1.0
tornado==6.0.3
tqdm==4.31.1
uiautomator2==0.3.3
urllib3==1.22
virtualenv==16.7.4
w3lib==1.20.0
wcwidth==0.1.7
websocket-client==0.57.0
websockets==7.0
weditor==0.2.3
Werkzeug==0.14.1
whichcraft==0.6.0
xlrd==1.1.0
xlwt==1.3.0
zipp==3.1.0
